+++ ceammc/ext/src/flow/flow_pass.h
@@ -29,8 +29,7 @@ public:
     void onList(const AtomList& l) override;
     void onAny(t_symbol* s, const AtomListView& lst) override;
 
-    void parseProperties() override;
-    bool processAnyProps(t_symbol* sel, const AtomListView& lst)  override;
+    bool processAnyProps(t_symbol* sel, const AtomListView& lst) override { return false; }
 };
 
 void setup_flow_pass();